// viasoft.analytics.userbehaviour-frontend/src/clients/user-behaviour-analytics/model/filterByGroupingKey.ts
/**
 * Viasoft.Analytics.UserBehaviour
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Domains } from './domains';


export interface FilterByGroupingKey { 
    appName?: string | null;
    appIdentifier?: string | null;
    domain?: Domains;
    licensingIdentifier?: string;
    accountName?: string | null;
    accountId?: string;
    value?: number;
    readonly domainName?: string | null;
}