// viasoft.analytics.userbehaviour-frontend/src/clients/user-behaviour-analytics/model/filterByGroupingOutput.ts
/**
 * Viasoft.Analytics.UserBehaviour
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FilterByGroupingKey } from './filterByGroupingKey';


export interface FilterByGroupingOutput { 
    values?: Array<FilterByGroupingKey> | null;
}